+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ jobs:
           bun-version: 1.1.42
 
       - name: Install dependencies
-        run: bun install
+        run: bun install --frozen-lockfile
 
       - name: Run tests
         run: bun test
