+++ .github/workflows/playwright.yml
@@ -24,7 +24,7 @@ jobs:
       run: bunx playwright install --with-deps
 
     - name: Run Playwright tests
-      run: bun run test
+      run: bun run test:e2e
 
     - uses: actions/upload-artifact@v4
       if: ${{ !cancelled() }}
