+++ e2e/fixtures/password-generator-fixture.ts
