+++ package.json
@@ -17,7 +17,8 @@
     "fix:all": "bun run typecheck && biome check --write --verbose",
     "fix:staged": "bun run typecheck && biome check --write --staged --verbose",
     "check:all": "run-p typecheck lint:check format:check",
-    "test": "playwright test"
+    "test": "bun test",
+    "test:e2e": "playwright test"
   },
   "scripts:comments": {
     "dev": "We are using tsx instead of bun because bun doesn't support all node APIs that Remix development server needs. See https://bun.sh/guides/ecosystem/remix. We need to exclude the config file from tsx watch because vite seems to be writing to those temporary files which causes watcher loop.",
